MarkdownComponent: Fix crash when rendering inline images
In ImageDrawable#draw, when the image loads, it calls md.layout().
However, MarkdownComponent#layout unfortunately calls reparse due
to the reasons outlined in the TODO comment. This causes the
ImageDrawable which invoked md.layout to be thrown away, however
execution of it's draw() method still continues, and throws an
UninitializedPropertyAccessException as image{X,Y} never get
assigned.

To fix this, the TODO comment in MarkdownComponent#layout was
addressed. ParagraphDrawable no longer mutates its drawable list,
and as such MarkdownComponent#layout no longer calls reparse(). Not
only does this fix the crash, it will also improve performance, as
a relayout doesn't also cause a reparse.